add remove_use()
Add an helper to remove the usage of a pseudo like kill_use() do
but without trying to kill (recursively!) the defining instruction
if the usage drop to zero.

This function is meant to be used when it is not yet safe to
kill instructions. If the usage drop to zero, nothing special is
done, the instruction becomes dead and will be eliminated later.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>