logical_instead_of_bitwise: silence some false positives
Ugh... The truth is that I have balls and balls of unpublished changes to
Smatch.  I've been going through and I noticed this.  Presumably it
silences a bunch of warnings.

I have a number of other proposed changes to this file but I need to go\
through and evaluate them before publishing.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>