project: create a way to generate smatch_data for a new project
Edit the Makefile to set CC to cgcc then run
smatch_scripts/build_generic_info.sh

There are a bunch of other small changes to make this slightly cleaner.
One thing is that now we prefer smatch_data/ in the CWD if it exists
instead of the one that ships with the smatch binary.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>