extra: handle integer overflows better for "int_foo += 4UL;"
Say you are adding an unsigned long to an integer on a 64 bit machine, then
it was checking if adding "(unsigned long)s64max + 4UL" can overflow.  It
can't so we would do the addition which gives a negative integer.  So then
it would say we now know that int_foo is s64min-(-2147483645).  What we
want it to say is that s32max + 4 can overflow so the value is
s32min-s32max still.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>