smdb.py: fix summary (it was totally broken before)
Wow.  I tested this, so I have no idea how this managed to be so utterly
broken.

Anyway the first "if not ret" test is completely wrong.  "ret" is always
known at that point.  Also if we add something to the beginning of the list
then the last bit wasn't getting appended properly.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>