extra: don't overwrite the implications for useless pointer limits
There was a function which said we limited a pointer to either zero or
valid.  We started out with an unknown value but with implications for
NULL and non-NULL.  So to limit it to NULL and valid is basically
useless.

Probably we shouldn't even store that useless information in the DB.
But anyway, I've filtered it out here.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>