kernel: ERR_PTR() basically only returns negatives and the occasional NULL
Say we don't know what a function returns, but we know that the result
is passed to ERR_PTR().  That probably means it's an ERR_PTR().  We
sometimes also pass NULLs to ERR_PTR() as well.  But we basically
shouldn't be passing positive values to ERR_PTR().

I've added a hack in check_kernel.c to force this behavior.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>