ranges: make sval_too_big() more accurate
Writing this changelog, I'm not totally sure we should even have the
sval_too_big() function.  What I'm doing is when we cast a range_list
from one type to the other, if the first range doesn't map to within
the second range then we just give up and say the result is the whole
range of the second smaller type.

The bug that I was trying to address was that casting a negative to
unsigned, such as u32 -> s32 was making Smatch just give up.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>