testsuite: remove unneeded './' before commands
Some testcase have their command specified as './<command name>'
but the './' part is unneeded as all commands are first prefixed
with '../' before being run.

Furthermore, the possible presence of these './' force the
use of 'basename' when filtering out the disabled commands.

Change this by stripping the './' where is was used.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>