testsuite: avoid fork+execing basename
Some testcase (the ones related to sparse-llvm) are disabled if
the needed support is not present. This is done by checking
the name of the command used by the testcase.

The previous possible presence of './' before the command meant
that the command was checked via the 'basename' command, which
need to ne fork+execed.
Since the './' have now been stripped from all command names
this is no more needed.

Change this by comparing directly the command name.

This speedup the testsuite by another 6%.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>