let handle_switches() also handle reverse logic
Some flags, for good or less good reasons, exist under two
different names: like a positive version and a negative one.
This is in addition of the 'no-' prefix.

The use case here is '-f[no-][un]signed-char'.

Make it easy and allow to specify that a given flag is the
negative version of another one so that '-fsigned-char' is
handled exactly as '-fno-unsigned-char'.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>