big-shift: don't take the modulo at expand time
Currently, at expansion time, shifts expressions with an
amount corresponding to undefined behaviour (larger or equal
than the type's width or negative) are 'reformed' by reducing
the amount modulo the width.
This correspond in fact to the run-time of several CPUs families
(x86[-64], arm64, mips, ...) but not all of them (arm, ppc, ...).

However, it is desirable for the front-end to not modify the
exacution model, the virtual one given by the C standard, but
also the effective one of the target machine.

Change this, but no more doing the reduction modulo the width
and leaving these expressions as-is (which leave the possibility
to do something else, maybe target-specific, at a latter stage).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>