big-shift: do not evaluate negative or over-sized shifts
During simplification, an instruction is evaluated and replaced
by its value when all its operands of an instruction are known
to be constant.

However, for shifts, not all amounts give a well defined result:
* when the amount is larger or equal to the type's width
* when the shift is negative
Thus performing this evaluation can possibly give a different
effect than what would happen if the instruction would be executed
on the target machine.

In one way, this doesn't really matter since it's undefined anyway.
But it is desirable for the simplification to be completly
deterministic.

So, don't perform this evaluation and leave these undefined operations
as they are (which still gives the possibility to handle them later,
depending on some f-options or maybe in a target-dependent way).

Note: this is largely what seems to be done by GCC.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>