force to 0 expressions which are erroneously non-constant
When an expression that needs to be constant but is, in fact,
not constant, sparse throws an error and leaves it as-is.
But some code makes the assumption that the expression is
constant and uses its value, with some random result.

One situation where this happens is in code like:
	switch (x) {
	case <some non-const expression>: ...

In this case, the linearization of the switch/case statement
will unconditionally use the value of the case expression
but the expression has no value.

One way to avoid this would be to add defensive checks each
time a value is retrieved but this is a lot of work and time
for no benefits.

So, change this by forcing the expression to be a constant
value of 0 just after the error message has been issued.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>