big-shift: do not simplify over-sized OP_ASR to zero
Currently, arithmetic right shifts with a shift count bigger
than the operand size are simplified to zero.

While this makes a lot of sense for LSR and SHL, for ASR it's
much less so.

Remove the simplification and let the back-end generate the code
for a non-immediate count (which is what GCC do).

Note: Some other options would be:
- reduce the shift count modulo the operand size, like it was
  done previously at expansion time and which correspond to the
  run-time behaviour of several CPUs families (x86[-64], arm64,
  mips, ...) but not all of them (arm, ppc, ...).
- truncate the count to N-1

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>