dissect: use built_in_ident() instead of MK_IDENT()
The motivation for this patch was to allow sparse to be
compiled with clang which doesn't like static initialization
of flexible array members which is used in MK_IDENT().

But also, there is no enough good justifications here for not
creating the identifiers the normal way with built_in_ident().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Acked-by: Oleg Nesterov <oleg@redhat.com>