teach sparse about -m{big,little}-endian
Some macros, structures definitions, ... are endianness specific.
This is generaly done with the help of some header files but these
headers often need information from the compiler which then need
to be endian-aware.

In order to not interpret things differently than the compiler
would, sparse thus also need to be endian-aware, even more so
given that sparse must cover all archs and their variants.

Give this information to sparse with two flags:
-mbig-endian/-mlittle-endian, and by default use
the endianness of the platform used to compile sparse.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Acked-by: Christopher Li <sparse@chrisli.org>