testsuite: respect command line's quotes & whitespaces
Currently the testsuite use 'eval echo $cmd' to expand
the name of the test file to be given on the command line.

This has the annoying consequence to go a bit too far in
the expansion of variables and to destroy any quotes and
whitespaces escaping that would have done.

Fix this by doing the eval later, when effectively executing
the command.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>