add helper: valid_type()
'bad_type' is used for ... bad types instead of NULL
in order to avoid later lots of null-checking.

Alas NULL is also often used and so we have to check
for NULL and 'bad_ctype' when checking a type's validity.

Add an helper: valid_type() to make the code a bit simpler.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>