use valid_type to avoid to warn twice on conditionals
When evaluating a conditional, the expression is first evaluated
and some further verifications and processing are done if
the returned type is not NULL.

However, the returned type can also be 'bad_ctype' and if it is
the case, the additional verifications will just give meaningless
additional warnings.

Fix this by using the new helper valid_type() instead of just
testing for a null ctype.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>