extra: don't call create new states for unknown values with param_set
If you start out with no state, then it doesn't really make sense to
set the state to unknown.  Unknown and no state are the same thing.

To be honest, I'm not sure why this is only applied to __in_fake_assign
and in_param_set().  Anyway, I've been running with this code for a while
and want to commit it so my private diffstat shrinks.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>