math: handle this (<too complicated math> & 0xff)
The recurse_cnt is used to give up on trying to parse too complicated
math.  There are one or two macros in the kernel that are pages and pages
of math and create out of memory issues.  But on the other hand, it's
really puzzling to see code like:

	foo = MACRO & 0xff;

and Smatch not be able to parse it.  And you might have to look really
closely at MACRO to see that actually it's a half a page of math when you
expand it completely.

Anyway, let's treat this as a special case because it's quite important.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>